server/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
sqlalchemy
aiosqlite
pycryptodomex
python-multipart
python-dotenv
jinja2
email-validator
